% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{add_bookmark}
\alias{add_bookmark}
\title{Add bookmark}
\usage{
add_bookmark(
  api_token = Sys.getenv("API_TOKEN"),
  url,
  title,
  extended = NULL,
  tags = NULL,
  date = NULL,
  replace = TRUE,
  shared = TRUE,
  toread = FALSE
)
}
\arguments{
\item{api_token}{(chr) api auth token.}

\item{url}{(chr) Url to add.}

\item{title}{(chr) Title of the bookmark.}

\item{extended}{(chr) Description of the item.}

\item{tags}{(chr vector) List of tags, up to 100 tags.}

\item{date}{(date) Creation date for the bookmark, to be formatted as 2020-02-23T12:32:45Z.}

\item{replace}{(bool) Replace an existing bookmark. Throws an error if set to false.}

\item{shared}{(bool) Make the bookmark public.}

\item{toread}{(bool) Flags the bookmark as unread. Default is FALSE.}
}
\value{
TRUE if successful, error message otherwise.
}
\description{
Add a bookmark to user collection.
}
