% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bool_to_yesno}
\alias{bool_to_yesno}
\title{Converts a boolean to yes/no string}
\usage{
bool_to_yesno(x)
}
\arguments{
\item{(BOOL)}{A boolean value}
}
\value{
(str) yes or no
}
\description{
Instead of booleans, pinboard.in API uses 'yes'/'no' strings.
}
\keyword{internal}
