% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{date_last_updated}
\alias{date_last_updated}
\title{Get time of last interaction}
\usage{
date_last_updated(api_token = Sys.getenv("API_TOKEN"))
}
\arguments{
\item{api_token}{(chr) api auth token.}
}
\description{
Returns the most recent time a bookmark was added, updated or deleted.
}
