% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{delete_bookmark}
\alias{delete_bookmark}
\title{Delete a bookmark}
\usage{
delete_bookmark(api_token = Sys.getenv("API_TOKEN"), url)
}
\arguments{
\item{api_token}{(chr) api auth token.}

\item{url}{(chr) Url to delete.}
}
\value{
TRUE if successful, error message otherwise.
}
\description{
Deletes a bookmark from user's collection.
}
