% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{delete_tag}
\alias{delete_tag}
\title{Delete a tag}
\usage{
delete_tag(api_token = Sys.getenv("API_TOKEN"), tag)
}
\arguments{
\item{api_token}{(chr) api auth token.}

\item{tag}{(chr) Tag to be deleted.}
}
\value{
TRUE if successful, error message otherwise.
}
\description{
Deletes a tag in user collection.
}
