% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{get_all_bookmarks}
\alias{get_all_bookmarks}
\title{Get all bookmarks}
\usage{
get_all_bookmarks(
  api_token = Sys.getenv("API_TOKEN"),
  tags = NULL,
  start = 0,
  n_results = NULL,
  from_date = NULL,
  to_date = NULL,
  meta = FALSE
)
}
\arguments{
\item{api_token}{(chr) api auth token.}

\item{tags}{(chr vector) List of tags, up to 100 tags.}

\item{start}{(int) Offset value. Default is 0.}

\item{n_results}{(int) Number of result to return. Default is all.}

\item{from_date}{(date) Return only bookmarks after this date. To be formatted as 2020-02-23T12:32:45Z.}

\item{to_date}{(date) Return only bookmarks before this date. To be formatted as 2020-02-23T12:32:45Z.}

\item{meta}{(BOOL) Include a change detection signature for each bookmark.}
}
\value{
TRUE if successful, error message otherwise.
}
\description{
Get all bookmarks in a user collection.
}
