% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{get_all_notes}
\alias{get_all_notes}
\title{Return all user notes}
\usage{
get_all_notes(api_token = Sys.getenv("API_TOKEN"))
}
\arguments{
\item{api_token}{api auth token}
}
\value{
A tibble.
}
\description{
Returns a dataframe of all notes created by user.
}
