% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_all_tags}
\alias{get_all_tags}
\title{Geta all tags}
\usage{
get_all_tags(api_token = Sys.getenv("API_TOKEN"))
}
\arguments{
\item{api_token}{(chr) api auth token.}
}
\value{
A tibble.
}
\description{
Returns a full list of the user's tags along with the number of times they were used.
}
