% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{get_note}
\alias{get_note}
\title{Return a specific user note}
\usage{
get_note(api_token = Sys.getenv("API_TOKEN"), id)
}
\arguments{
\item{api_token}{api auth token.}

\item{id}{Id of the note to be retrieved.}
}
\value{
A tibble.
}
\description{
Retrieve a specific note by id.
}
