% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{get_recent_bookmarks}
\alias{get_recent_bookmarks}
\title{Get recent bookmarks}
\usage{
get_recent_bookmarks(
  api_token = Sys.getenv("API_TOKEN"),
  tags = NULL,
  n_results = NULL
)
}
\arguments{
\item{api_token}{(chr) api auth token.}

\item{tags}{(chr vector) Filter up to three tags}

\item{n_results}{(int) Number of result to return. Default is all.}
}
\value{
TRUE if successful, error message otherwise.
}
\description{
Get user's recent bookmarks, filtered by tag.
}
