% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_secret}
\alias{get_secret}
\title{Get api secret}
\usage{
get_secret(api_token = Sys.getenv("API_TOKEN"))
}
\arguments{
\item{api_token}{api auth token.}
}
\value{
A string
}
\description{
Api secret allows to access user's private feeds.
}
