% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tag_recommendation}
\alias{get_tag_recommendation}
\title{Get tag recommendation for a URL}
\usage{
get_tag_recommendation(api_token = Sys.getenv("API_TOKEN"), url)
}
\arguments{
\item{api_token}{(chr) api auth token.}

\item{url}{(chr) URL address to get recommendations for.}
}
\value{
A list
}
\description{
Returns a list of popular tags and recommended tags for a given URL. Popular tags
  are tags used site-wide for the url; recommended tags are drawn from the user's own tags.
}
