% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{n_bookmarks_by_date}
\alias{n_bookmarks_by_date}
\title{Get number of boorkmarks by date}
\usage{
n_bookmarks_by_date(api_token = Sys.getenv("API_TOKEN"), tags = NULL)
}
\arguments{
\item{api_token}{(chr) api auth token.}

\item{tags}{(chr vector) Filter up to 3 tags.}
}
\value{
A tibble
}
\description{
Returns a tibble with the number of bookmarks created by date.
}
