% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_bool_res}
\alias{parse_bool_res}
\title{Parses a boolean API response}
\usage{
parse_bool_res(res)
}
\arguments{
\item{res}{An httr response}
}
\value{
TRUE if successful, error message otherwise.
}
\description{
Parses a boolean API response
}
\keyword{internal}
