% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinboard.R
\docType{package}
\name{pinboard}
\alias{pinboard}
\alias{res_format}
\title{A full-featured API wrapper for pinboard.in}
\format{
An object of class \code{character} of length 1.
}
\usage{
res_format
}
\description{
See documentation at \href{https://andodet.github.io/pinboard/}{https://andodet.github.io/pinboard/}
}
\keyword{datasets}
