% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{rename_tag}
\alias{rename_tag}
\title{Rename tag}
\usage{
rename_tag(api_token = Sys.getenv("API_TOKEN"), old, new)
}
\arguments{
\item{api_token}{api auth token.}

\item{old}{(chr) Tag to be renamed.}

\item{new}{(chr) New tag}
}
\value{
TRUE if successful, error message otherwise.
}
\description{
Rename a tag found in user's collection.
}
